requirements.txt
aniso8601==9.0.1
attrs==22.2.0
blessings==1.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colour-runner==0.1.1
cryptography==39.0.1
dis==1.0.1
flake8==6.0.0
Flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
packaging==23.0
pip-review==1.3.0
pluggy==1.0.0
prettytable==3.6.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.25
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
